IAppendFrameInputDataMonoBehaviourのテストコードを追加
