+++ Runtime/Input/FrameInputData/MonoBehaviour/AppendMouseInputData.cs
@@ -13,18 +13,26 @@ namespace Hinode
     /// <seealso cref="InputRecorderMonoBehaviour"/>
     /// <seealso cref="IAppendFrameInputDataMonoBehaviour"/>
     /// </summary>
-    public class AttachMouseInputData : IAppendFrameInputDataMonoBehaviour
+    public class AppendMouseInputData : IAppendFrameInputDataMonoBehaviour
     {
         #region override IAppendFrameInputDataMonoBehaviour
-        protected override void OnAwake(InputRecorder inputRecorder)
+        public override IFrameDataRecorder CreateInputData()
+        {
+            var btn = new MouseFrameInputData();
+            return btn;
+        }
+
+        protected override void OnAttached(InputRecorder inputRecorder)
         {
             if (inputRecorder.FrameDataRecorder is FrameInputData)
             {
                 MouseFrameInputData.RegistTypeToFrameInputData();
 
                 var frameInputData = inputRecorder.FrameDataRecorder as FrameInputData;
-                var touchInputData = new MouseFrameInputData();
-                frameInputData.AddChildRecorder(touchInputData);
+                frameInputData.RemoveChildRecorder(MouseFrameInputData.KEY_CHILD_INPUT_DATA_TYPE);
+
+                var inputData = CreateInputData();
+                frameInputData.AddChildRecorder(inputData);
             }
         }
         #endregion
